requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
tqdm==4.67.1
